initstateequil: make it a private template
